+++ HelloVulkan/Source/Camera.cpp
@@ -231,18 +231,18 @@ FrustumUBO Camera::GetFrustumUBO() const
 	return ubo;
 }
 
-glm::vec4 Camera::GetRayFromScreenToWorld(int screenPosX, int screenPosY) const
+glm::vec4 Camera::GetRayFromScreenToWorld(float screenPosX, float screenPosY) const
 {
 	glm::mat4 inverseViewMatrix_ = glm::inverse(viewMatrix_);
 	glm::vec4 rayClipSpace =
 		glm::vec4(
 			(static_cast<float>(screenPosX) / screenWidth_) * 2.f - 1.f,
 			(static_cast<float>(screenPosY) / screenHeight_) * 2.f - 1.f,
-			-1.0f,
-			1.0f);
+			-1.f,
+			1.f);
 	glm::vec4 rayViewSpace = inverseProjectionMatrix_ * rayClipSpace;
-	rayViewSpace.z = -1.0f;
-	rayViewSpace.w = 0.0f;
+	rayViewSpace.z = -1.f;
+	rayViewSpace.w = 0.f;
 
 	glm::vec4 rayWorldSpace = inverseViewMatrix_ * rayViewSpace;
 	rayWorldSpace = glm::normalize(rayWorldSpace);
